feat: add reth_collateral_target tracking, use v4 network balances ABI
1. Add reth_collateral_target storage slot for collateral buffer routing
2. Replace RocketNetworkBalancesV3 ABI with dedicated V4 ABI
3. Remove stale v3 ABI files
4. Add reth_collateral_target to initial state params and settings handler
