+++ protocols/substreams/ethereum-rocketpool/src/abi/mod.rs
@@ -2,5 +2,5 @@
 
 pub mod rocket_dao_protocol_proposal;
 pub mod rocket_deposit_pool_v4;
-pub mod rocket_network_balances_v3;
+pub mod rocket_network_balances_v4;
 pub mod rocket_token_reth;
