+++ protocols/substreams/ethereum-rocketpool/src/constants.rs
@@ -22,7 +22,7 @@ pub const ROCKET_DEPOSIT_POOL_ADDRESS_V4: [u8; 20] =
 /// These are EVM storage slots in RocketStorage (base slot 2 for uintStorage, 5 for boolStorage).
 /// The settings key format is: keccak256(keccak256("dao.protocol.setting.deposit") ++ settingPath)
 /// and the EVM slot is: keccak256(abi.encode(key, mapping_base_slot)).
-pub(crate) const ALL_STORAGE_SLOTS: [StorageLocation; 11] = [
+pub(crate) const ALL_STORAGE_SLOTS: [StorageLocation; 12] = [
     ROCKET_DEPOSIT_POOL_ETH_BALANCE_SLOT,
     DEPOSITS_ENABLED_SLOT,
     DEPOSIT_ASSIGN_ENABLED_SLOT,
@@ -34,6 +34,7 @@ pub(crate) const ALL_STORAGE_SLOTS: [StorageLocation; 11] = [
     MEGAPOOL_QUEUE_REQUESTED_TOTAL_SLOT,
     MEGAPOOL_QUEUE_INDEX_SLOT,
     EXPRESS_QUEUE_RATE_SLOT,
+    RETH_COLLATERAL_TARGET_SLOT,
 ];
 
 // ----------- Contract: Rocket Vault -----------
@@ -129,6 +130,19 @@ pub(crate) const MEGAPOOL_QUEUE_INDEX_SLOT: StorageLocation = StorageLocation {
 pub(crate) const EXPRESS_QUEUE_RATE_SLOT: StorageLocation = StorageLocation {
     name: "express_queue_rate",
     slot: hex!("76db7078bc37e9c3634c81dc384e741875c5d95ee6d5bcae0fb5d844d3189423"),
+    offset: 0,
+    number_of_bytes: 32,
+    signed: false,
+};
+
+// ----------- Contract: Rocket Storage (rETH collateral buffer) -----------
+/// Target rETH collateral rate as a fraction of total ETH, scaled by 1e18.
+/// E.g. 1e16 = 1% — the rETH contract should hold 1% of total protocol ETH as native ETH.
+/// Storage key: keccak256(keccak256("dao.protocol.setting.network") ++ "network.reth.collateral.target")
+/// in uintStorage (base slot 2).
+pub(crate) const RETH_COLLATERAL_TARGET_SLOT: StorageLocation = StorageLocation {
+    name: "reth_collateral_target",
+    slot: hex!("e1cd6c7fac18bc41fcd8660dbc3a1370373485f93fbccc910651118840f7c3a8"),
     offset: 0,
     number_of_bytes: 32,
     signed: false,
