+++ protocols/substreams/ethereum-rocketpool/substreams.yaml
@@ -38,6 +38,7 @@ params:
       "megapool_queue_requested_total": "0x0000000000000000000000000000000000000000000000000000000000000000",
       "megapool_queue_index": "0x0000000000000000000000000000000000000000000000000000000000000000",
       "express_queue_rate": "0x0000000000000000000000000000000000000000000000000000000000000004",
+      "reth_collateral_target": "0x000000000000000000000000000000000000000000000000002386f26fc10000",
       "total_eth": "0x00000000000000000000000000000000000000000000540e645ee4119f4d8b9e",
       "reth_supply": "0x00000000000000000000000000000000000000000000489a96a246a2e92bbbd1"
     }
